+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Fixed a bug in function intervene()
 
+* Steplist Creator Shiny App can now be used on smaller screens
+
 # epicmodel 0.1.1
 
 * Initial release on CRAN.